+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/ActiveDirectoryAuthenticationTimeoutRetryHelper.cs
@@ -6,7 +6,7 @@ using System;
 using System.ComponentModel;
 using System.Security.Cryptography;
 using System.Text;
-using Microsoft.Data.SqlClient.Parser;
+using Microsoft.Data.SqlClient.Parser.Login;
 
 namespace Microsoft.Data.SqlClient
 {
