+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Parser/Login/PreLoginOptions.cs
@@ -4,19 +4,17 @@
 
 #nullable enable
 
-namespace Microsoft.Data.SqlClient.Parser;
+namespace Microsoft.Data.SqlClient.Parser.Login;
 
-internal enum EncryptionOptions
+internal enum PreLoginOptions
 {
-    OFF,
-    ON,
-    NOT_SUP,
-    REQ,
-    LOGIN,
-    OPTIONS_MASK = 0x3f,
-
-    #if NETFRAMEWORK
-    CTAIP = 0x40,
-    CLIENT_CERT = 0x80,
-    #endif
+    VERSION,
+    ENCRYPT,
+    INSTANCE,
+    THREADID,
+    MARS,
+    TRACEID,
+    FEDAUTHREQUIRED,
+    NUMOPT,
+    LASTOPT = 255
 }
