+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Parser/SNILoadHandle.cs
@@ -8,6 +8,7 @@ using System.Runtime.InteropServices;
 using Interop.Windows.Sni;
 using Microsoft.Data.SqlClient.Internal;
 using Microsoft.Data.SqlClient.LocalDb;
+using Microsoft.Data.SqlClient.Parser.Login;
 
 namespace Microsoft.Data.SqlClient.Parser;
