+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Parser/Tokens/_SqlMetaData.cs
@@ -5,7 +5,7 @@
 using System;
 using System.Data;
 
-namespace Microsoft.Data.SqlClient.Parser;
+namespace Microsoft.Data.SqlClient.Parser.Tokens;
 
 internal sealed class _SqlMetaData : SqlMetaDataPriv
 {
